requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
pandas==2.1.4
openpyxl==3.1.2
sqlite3
